requirements.txt
pymupdf
sentence-transformers
faiss-cpu
chromadb
openai
python-dotenv
fastapi
uvicorn[standard]
pydantic
tqdm
numpy
pandas
